+++ src/automerge.js
@@ -57,8 +57,8 @@ function setField(state, objectId, key, value) {
     throw new TypeError('Map entries starting with underscore are not allowed: ' + key)
   }
 
-  if (typeof value === 'undefined') {
-    return deleteField(state, objectId, key)
+  if (!['object', 'boolean', 'number', 'string'].includes(typeof value)) {
+    throw new TypeError('Unsupported type of value: ' + (typeof value))
   } else if (isObject(value)) {
     const [newState, newId] = createNestedObjects(state, value)
     return makeOp(newState, { action: 'link', obj: objectId, key, value: newId })